# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: StorageHub API
  version: 1.0.0
  description: >
    REST API of the StorageHub self-service storage management system.
    All endpoints live under the `/api/v1` prefix. Everything except the
    explicitly public endpoints requires a JWT Bearer token (AD-5).
servers:
  - url: /api/v1
    description: Same-origin (Vite dev proxy / SPA-served deployment)

# Default security: JWT everywhere. Public endpoints (e.g. POST /auth/login)
# override this with `security: []` when they are added.
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [Tasks]
      summary: List tasks for staff task board
      description: >
        Guarded by ROLE_STAFF / ROLE_FACILITY_MANAGER / ROLE_SYSTEM_ADMINISTRATOR / ROLE_ADMIN.
        Returns list of tasks filtered by workDate, task type, status, or assigned staff.
      parameters:
        - name: workDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filter tasks by work date (YYYY-MM-DD).
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskType'
          description: Filter by task type.
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskStatus'
          description: Filter by task status.
        - name: assignedStaffId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter by assigned staff ID.
      responses:
        '200':
          description: List of tasks matching the filter.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      operationId: createTask
      tags: [Tasks]
      summary: Create a new operational task
      description: >
        Guarded by ROLE_STAFF / ROLE_FACILITY_MANAGER / ROLE_SYSTEM_ADMINISTRATOR / ROLE_ADMIN.
        Creates a task in the central task registry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [Tasks]
      summary: Get task details by ID
      description: >
        Guarded by ROLE_STAFF / ROLE_FACILITY_MANAGER / ROLE_SYSTEM_ADMINISTRATOR / ROLE_ADMIN.
        Returns details of a specific task.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Task ID.
      responses:
        '200':
          description: Task details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{id}/status:
    patch:
      operationId: updateTaskStatus
      tags: [Tasks]
      summary: Transition task status on the Kanban board
      description: >
        Guarded by ROLE_STAFF / ROLE_FACILITY_MANAGER / ROLE_SYSTEM_ADMINISTRATOR / ROLE_ADMIN.
        Updates task status (TODO -> IN_PROGRESS -> DONE) and logs an audit record in activity_logs.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Task ID to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskStatusRequest'
      responses:
        '200':
          description: Task status updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Blocked transition or missing prerequisite steps.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{id}/validate-reservation:
    post:
      operationId: validateCheckInReservation
      tags: [Tasks]
      summary: Validate reservation code and paid deposit for check-in task
      description: >
        Guarded by ROLE_STAFF / ROLE_FACILITY_MANAGER / ROLE_BUSINESS_OPS / ROLE_ADMIN / ROLE_SYSTEM_ADMINISTRATOR.
        Validates customer reservation code, checks that deposit is paid, and returns the 2-line financial breakdown.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Check-in Task ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateCheckInRequest'
      responses:
        '200':
          description: Check-in validation result with 2-line financial breakdown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInValidationDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Reservation is not in RESERVED status or unpaid deposit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{id}/activate-checkin:
    post:
      operationId: activateCheckIn
      tags: [Tasks]
      summary: Finalize check-in ritual, activate rental, and reveal access code
      description: >
        Guarded by ROLE_STAFF / ROLE_FACILITY_MANAGER / ROLE_ADMIN.
        Verifies that 100% rent is paid and latest contract is SIGNED. In a single atomic transaction:
        (1) Generates access code on Reservation, (2) transitions Reservation -> CHECKED_IN,
        (3) transitions Unit -> RENTED, (4) transitions Task -> DONE, and (5) sends notification to customer.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Check-in Task ID.
      responses:
        '200':
          description: Check-in successfully activated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInActivationDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Contract not signed, rent unpaid, or invalid state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT with role claim, 24h TTL, no refresh token (AD-5). Logout is the
        client dropping the token.

  # D6 reusable non-2xx bodies (deferred from 1.1, landed with the first
  # auth paths in 1.3): every usage renders the same Error envelope.
  responses:
    NotFound:
      description: >
        Resource was not found. Error envelope, code NOT_FOUND (404).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: >
        Missing, invalid or expired Bearer token - or bad credentials /
        inactive / locked account at login. Error envelope, code
        UNAUTHENTICATED (401). Carries `WWW-Authenticate: Bearer` on guarded
        paths. At login every failure branch shares one generic message.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >
        Authenticated but the token's role is not allowed on this endpoint
        (server-side permission matrix). Error envelope, code FORBIDDEN (403).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: >
        Request body failed validation (missing/invalid fields, duplicate
        email, password confirmation mismatch, terms not accepted). Error
        envelope, code VALIDATION_FAILED (400) with one fieldError per
        offending field.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      description: >
        The one error envelope for every non-2xx response, including security
        (401/403) and business-rule blocks (409). The FE renders toasts from
        this data and never parses the human message.
      required:
        - code
        - message
      additionalProperties: false
      properties:
        code:
          type: string
          description: >
            Machine-readable, stable, UPPER_SNAKE. Two layers share this field.
            Framework codes, issued by the backend error envelope for every
            non-2xx response including security 401/403: VALIDATION_FAILED
            (400, with fieldErrors), MALFORMED_REQUEST (400), UNAUTHENTICATED
            (401), FORBIDDEN (403), NOT_FOUND (404), METHOD_NOT_ALLOWED (405),
            UNSUPPORTED_MEDIA_TYPE (415), INTERNAL_ERROR (500, generic message
            only). Business codes are declared by each operation when its path
            lands in this file and travel in this same field; a business-rule
            block maps to 409.
          pattern: '^[A-Z][A-Z0-9_]*$'
          examples:
            - VALIDATION_FAILED
            - UNIT_UNAVAILABLE
            - BLOCKED_TRANSITION
        message:
          type: string
          description: Human-readable one-liner (what happened + consequence + next step).
        fieldErrors:
          type: array
          description: Present for validation failures; empty or absent otherwise.
          items:
            $ref: '#/components/schemas/FieldError'
        missingStep:
          type: string
          nullable: true
          description: Machine code for the missing closing step (FR-22 snap-back).
          example: 'RENT_PAYMENT_PENDING'
        stepLabel:
          type: string
          nullable: true
          description: Human-readable closing step name for rendering the snap-back toast.
          example: 'Rent payment is still pending on this check-in.'
    FieldError:
      type: object
      required:
        - field
        - message
      additionalProperties: false
      properties:
        field:
          type: string
          description: camelCase JSON path of the offending field.
        message:
          type: string

    # ------------------------------------------------------------------- list
...
    ReservationStatus:
      type: string
      description: Lifecycle status of a reservation.
      enum:
        - PENDING_PAYMENT
        - RESERVED
        - CHECKED_IN
        - CHECKOUT_REQUESTED
        - CLOSED
        - EXPIRED
        - CANCELLED
...
    TaskType:
      type: string
      enum:
        - CHECK_IN
        - CHECKOUT
        - CLEANING
        - SUPPORT
        - CONTRACT
      description: The 5 operational task types.
    TaskStatus:
      type: string
      enum:
        - TODO
        - IN_PROGRESS
        - DONE
      description: Status of the task on the Kanban board.
    TaskDto:
      type: object
      description: Representation of an operational task card on the Kanban board.
      required:
        - id
        - type
        - refCode
        - assignedStaffId
        - workDate
        - status
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          example: 1
        type:
          $ref: '#/components/schemas/TaskType'
        refCode:
          type: string
          nullable: true
          example: 'BK-1042'
        assignedStaffId:
          type: integer
          format: int64
          example: 2
        assignedStaffName:
          type: string
          nullable: true
          example: 'Minh Tran'
        workDate:
          type: string
          format: date
          example: '2026-10-03'
        status:
          $ref: '#/components/schemas/TaskStatus'
        unitCode:
          type: string
          nullable: true
          example: 'S-3'
        customerName:
          type: string
          nullable: true
          example: 'Lan Nguyen'
        dueDate:
          type: string
          format: date
          nullable: true
          example: '2026-10-03'
        timeSlot:
          type: string
          nullable: true
          example: 'Morning'
        title:
          type: string
          nullable: true
          example: 'Check-in BK-1042'
        description:
          type: string
          nullable: true
          example: 'Customer Lan Nguyen checking into Unit S-3'
    UpdateTaskStatusRequest:
      type: object
      required:
        - status
      additionalProperties: false
      properties:
        status:
          $ref: '#/components/schemas/TaskStatus'
        reason:
          type: string
          nullable: true
          description: Optional note or reason for the status transition.
    CreateTaskRequest:
      type: object
      required:
        - type
        - workDate
      additionalProperties: false
      properties:
        type:
          $ref: '#/components/schemas/TaskType'
        refCode:
          type: string
          nullable: true
        assignedStaffId:
          type: integer
          format: int64
          nullable: true
        workDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/TaskStatus'
        unitCode:
          type: string
          nullable: true
        customerName:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        timeSlot:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    ValidateCheckInRequest:
      type: object
      required:
        - reservationCode
      additionalProperties: false
      properties:
        reservationCode:
          type: string
          example: 'BK-1042'
    CheckInValidationDto:
      type: object
      required:
        - valid
        - taskId
      additionalProperties: false
      properties:
        valid:
          type: boolean
          example: true
        errorCode:
          type: string
          nullable: true
          example: 'DEPOSIT_UNPAID'
        errorMessage:
          type: string
          nullable: true
          example: 'Deposit payment has not been confirmed'
        taskId:
          type: integer
          format: int64
          example: 1
        reservationId:
          type: integer
          format: int64
          nullable: true
          example: 1
        reservationCode:
          type: string
          nullable: true
          example: 'BK-1042'
        customerName:
          type: string
          nullable: true
          example: 'Lan Nguyen'
        unitCode:
          type: string
          nullable: true
          example: 'S-3'
        depositAmountPaid:
          type: integer
          nullable: true
          example: 103500
        totalRentDue:
          type: integer
          nullable: true
          example: 1035000
        depositReceiptCode:
          type: string
          nullable: true
          example: 'RC-1727932800101'
        rentPaid:
          type: boolean
          example: false
        rentReceiptCode:
          type: string
          nullable: true
          example: 'RC-1727932900202'
        status:
          $ref: '#/components/schemas/ReservationStatus'
...
    CheckInActivationDto:
      type: object
      description: Response upon activating check-in and generating access code.
      required:
        - accessCode
        - reservationCode
        - reservationId
        - unitCode
        - reservationStatus
        - unitStatus
        - taskStatus
      additionalProperties: false
      properties:
        accessCode:
          type: string
          description: Sensitive access PIN/QR code.
          example: '482913'
        reservationCode:
          type: string
          example: 'BK-1042'
        reservationId:
          type: integer
          format: int64
          example: 1
        unitCode:
          type: string
          example: 'S-3'
        reservationStatus:
          $ref: '#/components/schemas/ReservationStatus'
        unitStatus:
          $ref: '#/components/schemas/UnitStatus'
        taskStatus:
          $ref: '#/components/schemas/TaskStatus'
...